Consistency, GUI, and improved method of input
Improved consistency of code and made the overridden "ProcessCmdKey"
handle input from the arrow keys on the game window instead of telling
"KeyDown" to handle them.
Began constructing the GUI for the client. Added an empty control that
will be used to represent an item, and made a functional cascading
inventory bar to hold the inventory items.